Increase ARC meta data limits
Historically the ARC has limited meta data usage to some fraction
of the total ARC size.  The rational behind this was to ensure that
the more important metadata blocks wouldn't result in all data blocks
(even frequently used ones) being evicted from the cache.

However, this concern doesn't appear to have been entirely warrented.
For meta data heavy workloads we want the entire ARC to be available
for these blocks to improve the cache hit rate.  Furthermore, in
practice the most important data blocks will still remain cached as
long as they are frequently accessed.

For reasons very similar to the above a 10% limit was also imposed
on the faction of the ARC which can be consumed by dnodes.  For many
workloads this is fine.  However, for metadata heavy workloads it can
result in a lot of needless churn in the ARC and wasted CPU time spent
evicting dnodes to honor this limit.  To avoid this it's reasonable
to allow dnodes to take a larger fraction of the ARC when needed and
then be evicted by the VFS which can do this more effeciently.

Thie change increases the zfs_arc_meta_limit_percent to 100% of the
ARC, and the zfs_arc_dnode_limit_percent to 75% of the meta limit.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Issue #9966